+++ advanced-custom-blocks.php
@@ -93,11 +93,11 @@ advanced_custom_blocks()
  */
 advanced_custom_blocks()
 	->register_component( new \AdvancedCustomBlocks\Views\Admin_Menu() )
-	->register_component( new \AdvancedCustomBlocks\Post_Types\Block() )
+	->register_component( new \AdvancedCustomBlocks\Post_Types\Block_Post() )
 	->register_component( new \AdvancedCustomBlocks\Blocks\Loader() )
 	->register_component( new \AdvancedCustomBlocks\Endpoints\Preview() );
 
 /**
- * Sometimes we need to do some things after the plugin is loaded, so call the PluginInterface::plugin_loaded().
+ * Sometimes we need to do some things after the plugin is loaded, so call the Plugin_Interface::plugin_loaded().
  */
 add_action( 'plugins_loaded', array( advanced_custom_blocks(), 'plugin_loaded' ) );
